requirements.txt
selenium
Appium-Python-Client
PyYaml
pytest
logging